test(cache): move `Entry` and `copyEntry` to the `helper.test.ts`
