+++ packages/cache/jest.config.ts
@@ -1,3 +1,5 @@
 import { jestConfig } from "@actions-kit/dev";
 
-export default jestConfig();
+export default jestConfig({
+  testMatch: ["**/*.test.ts", "!**/helper.test.ts"],
+});
