+++ packages/cache/src/cache.test.ts
@@ -6,28 +6,9 @@ import {
   jest,
   test,
 } from "@jest/globals";
-import { Minimatch } from "minimatch";
 import * as path from "path";
 import { restore, save } from "./cache";
-
-interface Dictionary<T> {
-  [key: string]: T;
-}
-
-interface Entry extends Dictionary<Entry> {}
-
-function copyEntry(src: Entry, dst: Entry, patterns: string[]): Entry {
-  if (patterns.length > 0) {
-    const mm = new Minimatch(patterns[0]);
-    for (const name in src) {
-      if (!mm.match(name)) continue;
-      if (dst[name] === undefined) dst[name] = {};
-      dst[name] = copyEntry(src[name], dst[name], patterns.slice(1));
-    }
-    return dst;
-  }
-  return src;
-}
+import { copyEntry, Entry } from "./helper.test";
 
 class Mock {
   static caches: Map<string, Entry> = new Map();
